requirements.txt
google-cloud-bigquery==3.29.0
python-dotenv==1.0.1